A program to Print words contained in a file in alphabetical order
